% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_R_help.R
\name{extract_help_as_html}
\alias{extract_help_as_html}
\title{huhu}
\usage{
extract_help_as_html(pkg, fn = NULL)
}
\description{
https://stackoverflow.com/questions/9192589/how-can-i-extract-text-from-rs-help-command
}
