% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrtools.addin.R
\docType{package}
\name{rrtools.addin-package}
\alias{rrtools.addin}
\alias{rrtools.addin-package}
\title{rrtools.addin: A graphical user interface for some of the main functions of the rrtools package}
\description{
The rrtools package provides functions to create an R package research compendium -
             a powerful tool to create, manage and share reproducible research. 
             rrtools.addin is a graphical interface to this package to make the start as easy as possible.
             It only provides one function to start the addin: rrtools_assistant().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nevrome/rrtools.addin}
  \item Report bugs at \url{https://github.com/nevrome/rrtools.addin/issues}
}

}
\author{
\strong{Maintainer}: Clemens Schmid \email{clemens@nevrome.de}

}
\keyword{internal}
